Seprated windows definitions and UI elements from linuxUI library
