+++ .gitignore
@@ -1,4 +1,5 @@
 # exclude all files in build folder
 /ldmicro/build/*
+/ldmicro/testMain.cpp
 /.vscode
